foundry_common::abi

Function encode_function_args

pub fn encode_function_args<I, S>(func: &Function, args: I) -> Result<Vec<u8>>
where
    I: IntoIterator<Item = S>,
    S: AsRef<str>,

Given a function and a vector of string arguments, it proceeds to convert the args to alloy [DynSolValue]s and then ABI encode them.