Module foundry_common::compile

Support for compiling [foundry_compilers::Project]

Structs

ContractInfo

How big the contract is and whether it is a dev contract where size limits can be neglected

ContractSources

Contract source code and bytecode.

ProjectCompiler

Builder type to configure how to compile a project.

SizeReport

Contracts with info about their size

SkipBuildFilters

Bundles multiple SkipBuildFilter into a single FileFilter

Enums

SkipBuildFilter

A filter that excludes matching contracts from the build

Constants

CONTRACT_SIZE_LIMIT 🔒

Functions

compile_from_source

Compiles an Etherscan source from metadata by creating a project. Returns the artifact_id, the file_id, and the bytecode

compile_target

Compiles target file path.

deployed_contract_size

Returns the size of the deployed contract

etherscan_project

Creates a [Project] from an Etherscan source.

is_match_exclude 🔒

Matches file only if the filter does not apply.