foundry_common::constants

Function is_known_system_sender

pub fn is_known_system_sender(sender: Address) -> bool

Returns whether the sender is a known L2 system sender that is the first tx in every block.

Transactions from these senders usually don’t have a any fee information OR set absurdly high fees that exceed the gas limit (See: https://github.com/foundry-rs/foundry/pull/10608)

See: ARBITRUM_SENDER, OPTIMISM_SYSTEM_ADDRESS and [Address::ZERO]