foundry_common

Module contracts

Commonly used contract types and functions.

Structs

BytecodeData

Subset of [CompactBytecode] excluding sourcemaps.

ContractData

Container for commonly used contract data.

ContractsByArtifact

Wrapper type that maps an artifact to a contract ABI and bytecode.

Constants

CALL_PROTECTION_BYTECODE_PREFIX 🔒

Libraries’ runtime code always starts with the following instruction: PUSH20 0x0000000000000000000000000000000000000000

Functions

bytecode_diff_score

Very simple fuzzy matching of contract bytecode.

compact_to_contract

Helper function to convert CompactContractBytecode ~> ContractBytecodeSome

count_different_bytes 🔒 ⚠

Returns the amount of different bytes between two slices.

find_matching_contract_artifact

Returns the target artifact given the path and name.

find_target_path

Returns the canonicalized target path for the given identifier.

get_contract_name

Returns contract name for a given contract identifier.

get_file_name

This returns the file name part, See get_contract_name

Type Aliases

ArtifactWithContractRef 🔒

ContractsByAddress

Wrapper type that maps an address to a contract identifier and contract ABI.