foundry_common
Function block_on_handle

pub fn block_on_handle<F: Future>(handle: &Handle, future: F) -> F::Output

Block on a future using the current tokio runtime on the current thread with the given handle.