Module foundry_common::fs

Contains various std::fs wrapper functions that also contain the target path in their errors

Functions

canonicalize_path

Canonicalize a path, returning an error if the path does not exist.

copy

Wrapper for std::fs::copy

create_dir

Wrapper for std::fs::create_dir

create_dir_all

Wrapper for std::fs::create_dir_all

create_file

Wrapper for File::create.

files_with_ext

Returns all files with the given extension under the root dir

json_files

Returns a list of absolute paths to all the json files under the root

normalize_path

Normalize a path, removing things like . and ...

open

Wrapper for std::fs::File::open

read

Wrapper for std::fs::read.

read_json_file

Reads the JSON file and deserialize it into the provided type.

read_link

Wrapper for std::fs::read_link.

read_to_string

Wrapper for std::fs::read_to_string.

remove_dir

Wrapper for std::fs::remove_dir

remove_dir_all

Wrapper for std::fs::remove_dir_all

remove_file

Wrapper for std::fs::remove_file.

write

Wrapper for std::fs::write

write_json_file

Writes the object as a JSON object.

Type Aliases

Result 🔒