foundry_common::io
Module shell

Utility functions for writing to stdout and stderr.

Originally from cargo.

Structs

Shell

An abstraction around console output that remembers preferences for output verbosity and color.

Enums

ColorChoice

Whether messages should use color output.

ShellOut 🔒

A Writeable object, either with or without color support.

TtyWidth

Terminal width.

Verbosity

The requested verbosity of output.

Statics

GLOBAL_SHELL 🔒

The global shell instance.

Functions

is_quiet

Returns whether the verbosity level is Verbosity::Quiet.

supports_color 🔒

verbosity

Returns the currently set verbosity.