foundry_common::preprocessor

Module deps

Structs

BytecodeDependency 🔒

Represents a single bytecode dependency.

BytecodeDependencyCollector 🔒

Walks over contract HIR and collects BytecodeDependencys and referenced contracts.

PreprocessorDependencies 🔒

Holds data about referenced source contracts and bytecode dependencies.

Enums

BytecodeDependencyKind 🔒

Represents a bytecode dependency kind.

Functions

handle_call_expr 🔒

Helper function to analyze and extract bytecode dependency from a given call expression.

named_arg 🔒

Helper function to extract value of a given named arg.

remove_bytecode_dependencies 🔒

Goes over all test/script files and replaces bytecode dependencies with cheatcode invocations.