Module foundry_common::provider::tower

Alloy-related tower middleware for retrying rate-limited requests and applying backoff.

Structs

RetryBackoffLayer

An Alloy Tower Layer that is responsible for retrying requests based on the error type. See [TransportError].

RetryBackoffService

An Alloy Tower Service that is responsible for retrying requests based on the error type. See [TransportError] and RateLimitRetryPolicy.

Functions

compute_unit_offset_in_secs 🔒

Calculates an offset in seconds by taking into account the number of currently queued requests, number of requests that were ahead in the queue when the request was first issued, the average cost a weighted request (heuristic), and the number of available compute units per seconds.