Module foundry_common::runtime_client

Wrap different providers

Structs

RuntimeClient

A provider that connects on first request allowing handling of different provider types at runtime

RuntimeClientBuilder

Builder for RuntimeClient

Enums

InnerClient 🔒

Enum representing a the client types supported by the runtime provider

RuntimeClientError

Error type for the runtime provider

Functions

build_auth 🔒

url_to_file_path 🔒