Module foundry_common::selectors

Support for handling/identifying selectors.

Structs

Artifact 🔒

ParsedSignatures

PossibleSigs

RawSelectorImportData

SelectorImportEffect 🔒

SelectorImportRequest 🔒

SelectorImportResponse

SelectorImportResult 🔒

SignEthClient

A client that can request API data from https://api.openchain.xyz

Enums

SelectorImportData

SelectorOrSig

SelectorType

Constants

MAX_TIMEDOUT_REQ 🔒

How many request can time out before we decide this is a spurious connection

REQ_TIMEOUT 🔒

The standard request timeout for API requests

SELECTOR_IMPORT_URL 🔒

SELECTOR_LOOKUP_URL 🔒

Functions

decode_calldata

Fetches all possible signatures and attempts to abi decode the calldata

decode_event_topic

Fetches an event signature given the 32 byte topic using https://api.openchain.xyz

decode_function_selector

Fetches a function signature given the selector https://api.openchain.xyz

decode_selector

Decodes the given function or event selector using https://api.openchain.xyz

decode_selectors

Decodes the given function or event selectors using https://api.openchain.xyz

import_selectors

uploads selectors to https://api.openchain.xyz using the given data

parse_signatures

Parses a list of tokens into function, event, and error signatures. Also handles JSON artifact files Ignores invalid tokens

pretty_calldata

Pretty print calldata and if available, fetch possible function signatures