foundry_common

Module serde_helpers

Misc Serde helpers for foundry crates.

Modules

duration

Enums

NumberOrHexU256

An enum that represents either a serde_json::Number integer, or a hex [U256].

Numeric

Helper type to parse both u64 and U256

NumericSeq

Helper type to deserialize sequence of numbers

Functions

deserialize_number

Deserializes a number from hex or int

deserialize_number_opt

Deserializes a number from hex or int, but optionally

deserialize_number_seq

Deserializes single integer params: 1, [1], ["0x01"]

from_int_or_hex

Deserializes the input into a U256, accepting both 0x-prefixed hex and decimal strings with arbitrary precision, defined by serde_json’s Number.

from_int_or_hex_opt

Deserializes the input into an Option<U256>, using from_int_or_hex to deserialize the inner value.