Trait foundry_common::traits::TestFunctionExt

pub trait TestFunctionExt {
    // Required methods
    fn is_invariant_test(&self) -> bool;
    fn is_fuzz_test(&self) -> bool;
    fn is_test(&self) -> bool;
    fn is_test_fail(&self) -> bool;
    fn is_setup(&self) -> bool;
}

Extension trait for Function.

Required Methods

fn is_invariant_test(&self) -> bool

Returns whether this function should be executed as invariant test.

fn is_fuzz_test(&self) -> bool

Returns whether this function should be executed as fuzz test.

fn is_test(&self) -> bool

Returns whether this function is a test.

fn is_test_fail(&self) -> bool

Returns whether this function is a test that should fail.

fn is_setup(&self) -> bool

Returns whether this function is a setUp function.