Trait foundry_common::types::ToAlloy

pub trait ToAlloy {
    type To;

    // Required method
    fn to_alloy(self) -> Self::To;
}

Conversion trait to easily convert from Ethers types to Alloy types.

Required Associated Types

type To

The corresponding Alloy type.

Required Methods

fn to_alloy(self) -> Self::To

Converts the Ethers type to the corresponding Alloy type.