Trait foundry_common::types::ToEthers

pub trait ToEthers {
    type To;

    // Required method
    fn to_ethers(self) -> Self::To;
}

Conversion trait to easily convert from Alloy types to Ethers types.

Required Associated Types

type To

The corresponding Ethers type.

Required Methods

fn to_ethers(self) -> Self::To

Converts the Alloy type to the corresponding Ethers type.