Crate foundry_common

foundry-common

Common utilities for building and using foundry’s tools.

Re-exports

pub use foundry_common_fmt as fmt;

pub use constants::*;

pub use contracts::*;

pub use traits::*;

pub use transactions::*;

Modules

abi

ABI related helper functions.

calc

Commonly used calculations.

compile

Support for compiling [foundry_compilers::Project]

constants

Commonly used constants.

contracts

Commonly used contract types and functions.

ens

ENS Name resolving utilities.

errors

Commonly used errors

evm

cli arguments for configuring the evm settings

fs

Contains various std::fs wrapper functions that also contain the target path in their errors

provider

Provider-related instantiation and usage utilities.

retry

Retry utilities.

selectors

Support for handling/identifying selectors.

serde_helpers

Misc Serde helpers for foundry crates.

shell

Helpers for printing to output

term

terminal utils

traits

Commonly used traits.

transactions

Wrappers for transactions.

utils 🔒

Uncategorised utilities.

Macros

cli_warn

Displays warnings on the cli

Functions

block_on

Block on a future using the current tokio runtime on the current thread.

block_on_handle

Block on a future using the current tokio runtime on the current thread with the given handle.

erc7201

Computes the storage slot as specified by ERC-7201, using the erc7201 formula ID.