foundry_common
Module abi

ABI related helper functions.

Functions

abi_decode_calldata

Decodes the calldata of the function

coerce_value

Helper function to coerce a value to a [DynSolValue] given a type string

encode_args

encode_function_args

Given a function and a vector of string arguments, it proceeds to convert the args to alloy [DynSolValue]s and then ABI encode them.

encode_function_args_packed

Given a function and a vector of string arguments, it proceeds to convert the args to alloy [DynSolValue]s and encode them using the packed encoding.

find_source

If the code at address is a proxy, recurse until we find the implementation.

get_event

Given an event signature string, it tries to parse it as a Event

get_func

Given a function signature string, it tries to parse it as a Function

get_func_etherscan

Given a function name, address, and args, tries to parse it as a Function by fetching the abi from etherscan. If the address is a proxy, fetches the ABI of the implementation contract.

get_indexed_event

Given an event without indexed parameters and a rawlog, it tries to return the event with the proper indexed parameters. Otherwise, it returns the original event.