foundry_common
Module compile

Support for compiling [foundry_compilers::Project]

Structs

ContractInfo

How big the contract is and whether it is a dev contract where size limits can be neglected

ProjectCompiler

Builder type to configure how to compile a project.

SizeReport

Contracts with info about their size

Constants

CONTRACT_INITCODE_SIZE_LIMIT 🔒

CONTRACT_RUNTIME_SIZE_LIMIT 🔒

Functions

compile_target

Compiles target file path.

contract_size 🔒

Returns the deployed or init size of the contract.

etherscan_project

Creates a [Project] from an Etherscan source.

with_compilation_reporter

Configures the reporter and runs the given closure.