Module foundry_common::compile

Support for compiling [foundry_compilers::Project]

Structs

ContractInfo

How big the contract is and whether it is a dev contract where size limits can be neglected

ProjectCompiler

Builder type to configure how to compile a project.

SizeReport

Contracts with info about their size

Constants

CONTRACT_SIZE_LIMIT 🔒

Functions

compile_target

Compiles target file path.

deployed_contract_size

Returns the size of the deployed contract

etherscan_project

Creates a [Project] from an Etherscan source.

with_compilation_reporter

Configures the reporter and runs the given closure.