foundry_common

Module constants

Commonly used constants.

Constants

ALCHEMY_FREE_TIER_CUPS

Alchemy free tier cups: https://docs.alchemy.com/reference/pricing-plans

ARBITRUM_SENDER

Arbitrum L1 sender address of the first transaction in every block. 0x00000000000000000000000000000000000a4b05

CONTRACT_MAX_SIZE

Maximum size in bytes (0x6000) that a contract can have.

DEFAULT_USER_AGENT

Default user agent set as the header for requests that don’t specify one.

DEV_CHAIN_ID

The dev chain-id, inherited from hardhat

NON_ARCHIVE_NODE_WARNING

Logged when an error is indicative that the user is trying to fork from a non-archive node.

OPTIMISM_SYSTEM_ADDRESS

The system address, the sender of the first transaction in every block: 0xdeaddeaddeaddeaddeaddeaddeaddeaddead0001

REQUEST_TIMEOUT

Default request timeout for http requests

SELECTOR_LEN

The first four bytes of the call data for a function call specifies the function to be called.

SYSTEM_TRANSACTION_TYPE

Transaction identifier of System transaction types

TYPE_BINDING_PREFIX

Prefix for auto-generated type bindings using forge bind-json.

Functions

is_impersonated_sig

is_impersonated_tx

is_known_system_sender

Returns whether the sender is a known L2 system sender that is the first tx in every block.