Module foundry_common::ens::EnsResolver

ENS Resolver interface.

contract EnsResolver {
    function addr(bytes32 node) view returns (address);
    function name(bytes32 node) view returns (string);
}

Structs

EnsResolverInstance

A EnsResolver instance.

addrCall

Returns the address associated with the specified node. Function with signature addr(bytes32) and selector 0x3b3b57de.

addrReturn

Returns the address associated with the specified node. Container type for the return parameters of the addr(bytes32) function.

nameCall

Returns the name associated with an ENS node, for reverse records. Function with signature name(bytes32) and selector 0x691f3431.

nameReturn

Returns the name associated with an ENS node, for reverse records. Container type for the return parameters of the name(bytes32) function.

Enums

EnsResolverCalls

Container for all the EnsResolver function calls.

Functions

new

Creates a new wrapper around an on-chain EnsResolver contract instance.