foundry_common::io

Module shell

Utility functions for writing to stdout and stderr.

Originally from cargo.

Structs

Shell

An abstraction around console output that remembers preferences for output verbosity and color.

Enums

ColorChoice

Whether messages should use color output.

OutputFormat

The requested output format.

OutputMode

The requested output mode.

ShellOut 🔒

A Writeable object, either with or without color support.

TtyWidth

Terminal width.

Statics

GLOBAL_SHELL 🔒

The global shell instance.

Functions

color_choice

Returns the current color choice.

is_json

Returns whether the output format is OutputFormat::Json.

is_quiet

Returns whether the output mode is OutputMode::Quiet.

set_verbosity

Set the verbosity level.

supports_color 🔒

verbosity

Returns the currently set verbosity level.

Type Aliases

Verbosity

The verbosity level.