Module foundry_common::provider

Provider-related instantiation and usage utilities.

Modules

retry

An utility trait for retrying requests based on the error type. See [TransportError].

runtime_transport

Runtime transport that connects on first request, which can take either of an HTTP, WebSocket, or IPC transport and supports retries based on CUPS logic.

tower

Alloy-related tower middleware for retrying rate-limited requests and applying backoff.

Structs

ProviderBuilder

Helper type to construct a RetryProvider

Functions

get_http_provider

Constructs a provider with a 100 millisecond interval poll if it’s a localhost URL (most likely an anvil or other dev node) and with the default, or 7 second otherwise.

resolve_path 🔒 Non-Windows

try_get_http_provider

Constructs a provider with a 100 millisecond interval poll if it’s a localhost URL (most likely an anvil or other dev node) and with the default, or 7 second otherwise.

Type Aliases

RetryProvider

Helper type alias for a retry provider

RetryProviderWithSigner

Helper type alias for a retry provider with a signer