foundry_common

Module selectors

Support for handling/identifying selectors.

Structs

Artifact 🔒

OpenChainClient

A client that can request API data from OpenChain.

ParsedSignatures

PossibleSigs

RawSelectorImportData

SelectorImportEffect 🔒

SelectorImportRequest 🔒

SelectorImportResponse

SelectorImportResult 🔒

Enums

SelectorImportData

SelectorOrSig

SelectorType

The type of selector fetched from OpenChain.

Constants

BASE_URL 🔒

MAX_TIMEDOUT_REQ 🔒

How many request can time out before we decide this is a spurious connection.

REQ_TIMEOUT 🔒

The standard request timeout for API requests.

SELECTOR_IMPORT_URL 🔒

SELECTOR_LOOKUP_URL 🔒

Functions

decode_calldata

Fetches all possible signatures and attempts to abi decode the calldata using OpenChain.

decode_event_topic

Fetches an event signature given the 32 byte topic using OpenChain.

decode_function_selector

Fetches a function signature given the selector using OpenChain.

decode_selector

Decodes the given function or event selector using OpenChain.

decode_selectors

Decodes the given function or event selectors using OpenChain.

import_selectors

uploads selectors to OpenChain using the given data

parse_signatures

Parses a list of tokens into function, event, and error signatures. Also handles JSON artifact files Ignores invalid tokens

pretty_calldata

Pretty print calldata and if available, fetch possible function signatures.