Module foundry_common::shell

Helpers for printing to output

Structs

InstallError

Error indicating that set_hook was unable to install the provided ErrorHook

Shell

An abstraction around console output that also considers verbosity

WriteShellOut

A guarded shell output type

Enums

ShellOut

A Writeable object, either with or without color support

Verbosity

The requested verbosity of output.

Statics

SHELL 🔒

Stores the configured shell for the duration of the program

Traits

ShellWrite

Helper trait for custom shell output

Functions

eprintln

Prints the given message to the shell

print_json

Prints the given message to the shell

println

Prints the given message to the shell

set_shell

Install the provided shell

verbosity

Returns the configured verbosity

with_shell

Runs the given closure with the current shell, or default shell if none was set