foundry_common::traits

Trait TestFunctionExt

pub trait TestFunctionExt {
    // Provided methods
    fn test_function_kind(&self) -> TestFunctionKind { ... }
    fn is_setup(&self) -> bool { ... }
    fn is_any_test(&self) -> bool { ... }
    fn is_any_test_fail(&self) -> bool { ... }
    fn is_unit_test(&self) -> bool { ... }
    fn is_before_test_setup(&self) -> bool { ... }
    fn is_fuzz_test(&self) -> bool { ... }
    fn is_invariant_test(&self) -> bool { ... }
    fn is_after_invariant(&self) -> bool { ... }
    fn is_fixture(&self) -> bool { ... }
    fn is_reserved(&self) -> bool { ... }
}

Extension trait for Function.

Provided Methods

fn test_function_kind(&self) -> TestFunctionKind

Returns the kind of test function.

fn is_setup(&self) -> bool

Returns true if this function is a setUp function.

fn is_any_test(&self) -> bool

Returns true if this function is a unit, fuzz, or invariant test.

fn is_any_test_fail(&self) -> bool

Returns true if this function is a test that should fail.

fn is_unit_test(&self) -> bool

Returns true if this function is a unit test.

fn is_before_test_setup(&self) -> bool

Returns true if this function is a beforeTestSetup function.

fn is_fuzz_test(&self) -> bool

Returns true if this function is a fuzz test.

fn is_invariant_test(&self) -> bool

Returns true if this function is an invariant test.

fn is_after_invariant(&self) -> bool

Returns true if this function is an afterInvariant function.

fn is_fixture(&self) -> bool

Returns true if this function is a fixture function.

fn is_reserved(&self) -> bool

Returns true if this function is test reserved function.